+++ config/hyde-layouts-manager.php
@@ -86,9 +86,9 @@ return [
             ]
         ],
         //'your-custom-theme' => [
-        //'app' => 'hyde-layouts-manager::layouts.your-custom-theme.app',
-        //'styles' => 'vendor/your-custom-theme/css/melasistema/app.css',
-        //'scripts' => 'vendor/your-custom-theme/js/melasistema/app.js',
+            //'app' => 'hyde-layouts-manager::layouts.your-custom-theme.app',
+            //'styles' => 'vendor/your-custom-theme/css/melasistema/app.css',
+            //'scripts' => 'vendor/your-custom-theme/js/melasistema/app.js',
         //],
     ],
 
@@ -104,39 +104,51 @@ return [
         'hero' => [
             'view' => 'hyde-layouts-manager::components.hero',
             'default' => [
-                'bgColor' => 'bg-white',
-                'darkBgColor' => 'dark:bg-gray-900',
-                'padding' => 'py-16',
-                'textColor' => 'text-gray-900',
-                'darkTextColor' => 'dark:text-white',
-                'headingText' => 'HydePHP Layouts Manager',
-                'subHeadingText' => 'Manage your layouts and reusable components with ease.',
-                'align' => 'center',
-                'primaryButton' => [
-                    'text' => '<svg class="w-6 h-6" xmlns="http://www.w3.org/2000/svg" viewBox="0 0 98 96"><path fill-rule="evenodd" clip-rule="evenodd" d="M48.854 0C21.839 0 0 22 0 49.217c0 21.756 13.993 40.172 33.405 46.69 2.427.49 3.316-1.059 3.316-2.362 0-1.141-.08-5.052-.08-9.127-13.59 2.934-16.42-5.867-16.42-5.867-2.184-5.704-5.42-7.17-5.42-7.17-4.448-3.015.324-3.015.324-3.015 4.934.326 7.523 5.052 7.523 5.052 4.367 7.496 11.404 5.378 14.235 4.074.404-3.178 1.699-5.378 3.074-6.6-10.839-1.141-22.243-5.378-22.243-24.283 0-5.378 1.94-9.778 5.014-13.2-.485-1.222-2.184-6.275.486-13.038 0 0 4.125-1.304 13.426 5.052a46.97 46.97 0 0 1 12.214-1.63c4.125 0 8.33.571 12.213 1.63 9.302-6.356 13.427-5.052 13.427-5.052 2.67 6.763.97 11.816.485 13.038 3.155 3.422 5.015 7.822 5.015 13.2 0 18.905-11.404 23.06-22.324 24.283 1.78 1.548 3.316 4.481 3.316 9.126 0 6.6-.08 11.897-.08 13.526 0 1.304.89 2.853 3.316 2.364 19.412-6.52 33.405-24.935 33.405-46.691C97.707 22 75.788 0 48.854 0z" fill="#fff"/></svg> <span>GitHub</span>',
-                    'link' => 'https://github.com/melasistema/hydephp-layouts-manager',
-                    'bgColor' => 'bg-blue-700',
-                    'textColor' => 'text-white',
-                    'darkBgColor' => 'dark:bg-gray-900',
-                    'darkTextColor' => 'dark:text-white',
+                'layout' => [
+                    'showSubHeadingText' => true,  // New flag to control visibility of the subheading
+                    'showPrimaryButton' => true, // New flag to control visibility of the primary button
+                    'showSecondaryButton' => true, // New flag to control visibility of the secondary button
                 ],
-                'secondaryButton' => [
-                    'text' => 'About Me',
-                    'link' => 'https://github.com/melasistema',
+                'settings' => [
                     'bgColor' => 'bg-white',
-                    'textColor' => 'text-indigo-500',
                     'darkBgColor' => 'dark:bg-gray-900',
+                    'padding' => 'py-16',
+                    'textColor' => 'text-gray-900',
                     'darkTextColor' => 'dark:text-white',
-                ],
+                    'headingText' => 'HydePHP Layouts Manager',
+                    'subHeadingText' => 'Manage your layouts and reusable components with ease.',
+                    'align' => 'center',
+                    'primaryButton' => [
+                        'text' => '<svg class="w-6 h-6" xmlns="http://www.w3.org/2000/svg" viewBox="0 0 98 96"><path fill-rule="evenodd" clip-rule="evenodd" d="M48.854 0C21.839 0 0 22 0 49.217c0 21.756 13.993 40.172 33.405 46.69 2.427.49 3.316-1.059 3.316-2.362 0-1.141-.08-5.052-.08-9.127-13.59 2.934-16.42-5.867-16.42-5.867-2.184-5.704-5.42-7.17-5.42-7.17-4.448-3.015.324-3.015.324-3.015 4.934.326 7.523 5.052 7.523 5.052 4.367 7.496 11.404 5.378 14.235 4.074.404-3.178 1.699-5.378 3.074-6.6-10.839-1.141-22.243-5.378-22.243-24.283 0-5.378 1.94-9.778 5.014-13.2-.485-1.222-2.184-6.275.486-13.038 0 0 4.125-1.304 13.426 5.052a46.97 46.97 0 0 1 12.214-1.63c4.125 0 8.33.571 12.213 1.63 9.302-6.356 13.427-5.052 13.427-5.052 2.67 6.763.97 11.816.485 13.038 3.155 3.422 5.015 7.822 5.015 13.2 0 18.905-11.404 23.06-22.324 24.283 1.78 1.548 3.316 4.481 3.316 9.126 0 6.6-.08 11.897-.08 13.526 0 1.304.89 2.853 3.316 2.364 19.412-6.52 33.405-24.935 33.405-46.691C97.707 22 75.788 0 48.854 0z" fill="#fff"/></svg> <span>GitHub</span>',
+                        'link' => 'https://github.com/melasistema/hydephp-layouts-manager',
+                        'bgColor' => 'bg-blue-700',
+                        'textColor' => 'text-white',
+                        'darkBgColor' => 'dark:bg-blue-900',
+                        'darkTextColor' => 'dark:text-white',
+                    ],
+                    'secondaryButton' => [
+                        'text' => 'About Me',
+                        'link' => 'https://github.com/melasistema',
+                        'bgColor' => 'bg-white',
+                        'textColor' => 'text-indigo-500',
+                        'darkBgColor' => 'dark:bg-gray-900',
+                        'darkTextColor' => 'dark:text-white',
+                    ],
+                ]
             ],
         ],
         'accordion' => [
             'view' => 'hyde-layouts-manager::components.accordion',
             'default' => [
-                'bgColor' => 'bg-gray-500',          // Default background color
-                'darkBgColor' => 'dark:bg-gray-900',  // Default dark mode background color
-                'textColor' => 'text-white',          // Default text color
-                'darkTextColor' => 'dark:text-white', // Default dark mode text color
+                'layout' => [
+                    // Add any additional layout options here in the future
+                ],
+                'settings' => [
+                    'bgColor' => 'bg-white',          // Default background color
+                    'darkBgColor' => 'dark:bg-gray-900',  // Default dark mode background color
+                    'textColor' => 'text-gray-900',          // Default text color
+                    'darkTextColor' => 'dark:text-white', // Default dark mode text color
+                ],
                 'items' => [                          // Default accordion items
                     [
                         'title' => '🚀 What is HydePHP Layouts Manager?',
@@ -200,12 +212,47 @@ return [
         'carousel' => [
             'view' => 'hyde-layouts-manager::components.carousel',
             'default' => [
+                'layout' => [
+                    'showIndicators' => true,  // New flag to control visibility of the indicators
+                    'showControls' => true,    // New flag to control visibility of the controls
+                    'rounded' => false,         // New flag to control rounded corners
+                ],
                 'images' => [
                     'carousel/example/carousel-1.svg',
                     'carousel/example/carousel-2.svg',
                     'carousel/example/carousel-3.svg',
                     'carousel/example/carousel-4.svg',
                     'carousel/example/carousel-5.svg',
+                ],
+            ],
+        ],
+        'gallery' => [
+            'view' => 'hyde-layouts-manager::components.gallery',
+            'default' => [
+                'layout' => [
+                    // Default grid column counts for each breakpoint
+                    'cols' => [
+                        'default' => 3, // Default columns for 'lg' and 'xl' screens
+                        'sm' => '',    // Small screens
+                        'md' => '',    // Medium screens
+                        'lg' => '',    // Large screens
+                        'xl' => '',    // Extra-large screens
+                    ],
+                    'gap' => 'gap-4',  // Default gap between items
+                ],
+                'images' => [
+                    'gallery/hyde/01-hyde-logo-cut.png',
+                    'gallery/hyde/02-hyde-logo-cut.png',
+                    'gallery/hyde/03-hyde-logo-cut.png',
+                    'gallery/hyde/04-hyde-logo-cut.png',
+                    'gallery/hyde/05-hyde-logo-cut.png',
+                    'gallery/hyde/06-hyde-logo-cut.png',
+                    'gallery/hyde/07-hyde-logo-cut.png',
+                    'gallery/hyde/08-hyde-logo-cut.png',
+                    'gallery/hyde/09-hyde-logo-cut.png',
+                    'gallery/hyde/10-hyde-logo-cut.png',
+                    'gallery/hyde/11-hyde-logo-cut.png',
+                    'gallery/hyde/12-hyde-logo-cut.png',
                 ],
             ],
         ],